+++ src/framework/effects/BiquadNotch.java
@@ -3,19 +3,19 @@ package framework.effects;
 public class BiquadNotch extends Biquad {
 
 	@Override
-	public void update(double frequency, double filterQ) {
+	public void update(float frequency, float filterQ) {
 		frequency = normalizeAndClip(frequency);
 		if (frequency > 0 && frequency < 1) {
 			if (filterQ > 0) {
-				double w0 = Math.PI * frequency;
-				double alpha = Math.sin(w0) / (2 * filterQ);
-				double k = Math.cos(w0);
-				double b0 = 1;
-				double b1 = -2 * k;
-				double b2 = 1;
-				double a0 = 1 + alpha;
-				double a1 = -2 * k;
-				double a2 = 1 - alpha;
+				float w0 = (float) Math.PI * frequency;
+				float alpha = (float) Math.sin(w0) / (2 * filterQ);
+				float k = (float) Math.cos(w0);
+				float b0 = 1;
+				float b1 = -2 * k;
+				float b2 = 1;
+				float a0 = 1 + alpha;
+				float a1 = -2 * k;
+				float a2 = 1 - alpha;
 				setCoefficients(b0, b1, b2, a0, a1, a2);
 			} else {
 				// When filterQ = 0, the z-transform gives lim_{filterQ -> 0} = 0.
