+++ src/framework/effects/Gain.java
@@ -2,18 +2,18 @@ package framework.effects;
 
 public class Gain implements EffectsProcessor {
 	
-	public double[] process(double[] buffer, double gain) {
-		double[] processedBuffer = new double[buffer.length];
+	public float[] process(float[] buffer, float gain) {
+		float[] processedBuffer = new float[buffer.length];
 		for (int i = 0; i < buffer.length; i++) {
 			processedBuffer[i] = gain * buffer[i];
 		}
 		return processedBuffer;
 	}
 
-	public double[] process(double[] buffer, double[] gains) {
-		double[] processedBuffer = new double[buffer.length];
-		double gainIndex = 0;
-		double gainIncrement = (double) gains.length / buffer.length;
+	public float[] process(float[] buffer, float[] gains) {
+		float[] processedBuffer = new float[buffer.length];
+		float gainIndex = 0;
+		float gainIncrement = (float) gains.length / buffer.length;
 		for (int i = 0; i < buffer.length; i++) {
 			processedBuffer[i] = gains[(int) gainIndex] * buffer[i];
 			gainIndex += gainIncrement;
