+++ src/framework/effects/Loop.java
@@ -4,14 +4,14 @@ public class Loop implements EffectsProcessor {
 
 	// TODO implement variable parameters
 
-	public double[] process(double[] buffer, int start, int end) {
+	public float[] process(float[] buffer, int start, int end) {
 		int count = buffer.length / (end - start);
 		return process(buffer, start, end, count); // try to preserve the duration
 	}
 
-	public double[] process(double[] buffer, int start, int end, int count) {
+	public float[] process(float[] buffer, int start, int end, int count) {
 		int samples = (end - start) * count;
-		double[] processedBuffer = new double[samples];
+		float[] processedBuffer = new float[samples];
 		for (int i = 0, j = 0; i < samples; i++, j++) {
 			processedBuffer[i] = buffer[j % end + start];
 		}
