+++ src/framework/effects/Splice.java
@@ -4,10 +4,10 @@ public class Splice implements EffectsProcessor {
 	
 	// TODO implement cross fades
 	
-	public double[] process(double[]... buffers) {
+	public float[] process(float[]... buffers) {
 		int samples = 0;
-		for (double[] buffer : buffers) samples += buffer.length;
-		double[] processedBuffer = new double[samples];
+		for (float[] buffer : buffers) samples += buffer.length;
+		float[] processedBuffer = new float[samples];
 		for (int i = 0, j = 0, k = 0; i < samples; i++, k++) {
 			processedBuffer[i] = buffers[j][k];
 			if (k == buffers[j].length - 1) {
