+++ src/framework/generators/BreakpointFunction.java
@@ -5,32 +5,32 @@ import framework.utilities.PlotUtility;
 public class BreakpointFunction {
 	
 	private final int length;
-	public final double[] breakpoints;
-	public final double[] durations;
-	public final double[] weights;
+	public final float[] breakpoints;
+	public final float[] durations;
+	public final float[] weights;
 	
 	// TODO add more constructors
 	// TODO Exponential functions
 	
-	public BreakpointFunction(int length, double[] breakpoints) {
+	public BreakpointFunction(int length, float[] breakpoints) {
 		this.length = length; // between each breakpoint
 		this.breakpoints = breakpoints;
 		this.durations = null;
 		this.weights = null;
 	}
 	
-	public BreakpointFunction(int length, double[] breakpoints, double[] durations, double[] weights) {
+	public BreakpointFunction(int length, float[] breakpoints, float[] durations, float[] weights) {
 		this.length = length; // between each breakpoint
 		this.breakpoints = breakpoints;
 		this.durations = durations;
 		this.weights = weights;
 	}
 	
-	public double[] get() {
-		double[] array = new double[length * (breakpoints.length - 1)];
+	public float[] get() {
+		float[] array = new float[length * (breakpoints.length - 1)];
 		int index = 0;
-		double value = breakpoints[0];
-		double increment = 0;
+		float value = breakpoints[0];
+		float increment = 0;
 		for (int i = 0; i < breakpoints.length - 1; i++) {
 			for (int j = 0; j < length; j++) {
 				if (i != breakpoints.length - 2) {
