+++ src/framework/generators/NoiseGenerator.java
@@ -16,21 +16,21 @@ public class NoiseGenerator {
 		this.wavetable = wavetable;
 	}
 
-	public double[] get(int duration, double amplitude) {
+	public float[] get(int duration, float amplitude) {
 		int samples = duration * Settings.samplingRate / 1000;
-		double[] buffer = new double[samples];
+		float[] buffer = new float[samples];
 		for (int i = 0; i < samples; i++) {
 			buffer[i] = amplitude * wavetable.getSample();
 		}
 		return buffer;
 	}
 
-	public double[] get(int duration, double[] envelope) {
+	public float[] get(int duration, float[] envelope) {
 		int samples = duration * Settings.samplingRate / 1000;
-		double oscAmp = 0;
-		double envelopeIndex = 0;
-		double envelopeIncrement = (double) envelope.length / samples;
-		double[] buffer = new double[samples];
+		float oscAmp = 0;
+		float envelopeIndex = 0;
+		float envelopeIncrement = (float) envelope.length / samples;
+		float[] buffer = new float[samples];
 		for (int i = 0; i < samples; i++) {
 			oscAmp = envelope[(int) envelopeIndex];
 			buffer[i] = oscAmp * wavetable.getSample();
